Minor HTML change with sizing, CSS colors, text style, and media query
I changed the images of the UX skills and changed their sizes to fit better on the screen. I did most of the styling of CSS on this push as well. Lastly, I finalized the media query at the end. You will notice that there is a version of the query that is commented out. That is where I tried to implement the grid-template-areas but It would not work how I wanted it to. I was still able to get the RWD to work effectively using flex and block ... but still included the other version to show that I tried.